Refactor: calcCart 함수 분리 함수 모듈화.
