+++ src/basic/main.basic.js
@@ -11,20 +11,6 @@ import Selector from './components/Selector';
 import SelectorContainer from './components/SelectorContainer';
 import StockStatus from './components/StockStatus';
 import {
-  PRODUCT_ONE,
-  PRODUCT_TWO,
-  PRODUCT_THREE,
-  PRODUCT_FOUR,
-  PRODUCT_FIVE,
-  TEN_PERCENT,
-  FIFTEEN_PERCENT,
-  TWENTY_PERCENT,
-  FIVE_PERCENT,
-  TWENTY_FIVE_PERCENT,
-  DISCOUNT_STANDARD_COUNT,
-  VOLUME_ORDER_COUNT,
-} from './constants/enum';
-import {
   renderBonusPoints,
   renderCartSummaryDetails,
   renderDiscountInfo,
@@ -33,6 +19,7 @@ import {
 } from './render';
 import store from './store';
 import { getStockInfoMessage } from './utils';
+import { applyAdditionalDiscounts, getCartSummary } from './utils/cart';
 
 const { productStore } = store;
 
@@ -42,64 +29,6 @@ const $selector = Selector();
 const $addButton = AddButton();
 const $cartItems = CartItems();
 const $stockStatus = StockStatus();
-
-// 1. 가격·수량 계산 및 할인 정보 추출
-function getCartSummary(cartItemList, productStore) {
-  let itemCount = 0,
-    subTotal = 0,
-    totalAmount = 0;
-  const itemDiscounts = [];
-
-  cartItemList.forEach((cartItem) => {
-    const product = productStore.getProductById(cartItem.id);
-    const quantity = parseInt(cartItem.querySelector('.quantity-number').textContent);
-    const itemTotal = product.value * quantity;
-    let discount = 0;
-
-    itemCount += quantity;
-    subTotal += itemTotal;
-
-    if (quantity >= DISCOUNT_STANDARD_COUNT) {
-      if (product.id === PRODUCT_ONE) discount = TEN_PERCENT;
-      else if (product.id === PRODUCT_TWO) discount = FIFTEEN_PERCENT;
-      else if (product.id === PRODUCT_THREE) discount = TWENTY_PERCENT;
-      else if (product.id === PRODUCT_FOUR) discount = FIVE_PERCENT;
-      else if (product.id === PRODUCT_FIVE) discount = TWENTY_FIVE_PERCENT;
-      if (discount > 0) itemDiscounts.push({ name: product.name, discount: discount * 100 });
-    }
-
-    totalAmount += itemTotal * (1 - discount);
-  });
-
-  return { itemCount, subTotal, totalAmount, itemDiscounts };
-}
-
-// 2. 대량/화요일 할인 적용 및 화요일 배너 표시
-function applyAdditionalDiscounts({ subTotal, totalAmount, itemCount, isTuesday }) {
-  let discountRate = 0,
-    finalAmount = totalAmount,
-    originalTotal = subTotal;
-
-  if (itemCount >= VOLUME_ORDER_COUNT) {
-    finalAmount = (subTotal * 75) / 100;
-    discountRate = TWENTY_FIVE_PERCENT;
-  } else {
-    discountRate = (subTotal - totalAmount) / subTotal;
-  }
-
-  // 화요일 배너 표시 관련 변수는 여기서 못 처리하므로 돌아온 후 calcCart에서 처리
-
-  if (isTuesday && finalAmount > 0) {
-    finalAmount = (finalAmount * 90) / 100;
-    discountRate = 1 - finalAmount / originalTotal;
-  }
-
-  return { finalAmount, discountRate, originalTotal };
-}
-
-function updateStockInfo(productList) {
-  $stockStatus.textContent = getStockInfoMessage(productList);
-}
 
 // 장바구니 계산 컨트롤러
 function calcCart() {
@@ -114,7 +43,7 @@ function calcCart() {
     productStore,
   );
 
-  let { finalAmount, discountRate, originalTotal } = applyAdditionalDiscounts({
+  const { finalAmount, discountRate, originalTotal } = applyAdditionalDiscounts({
     subTotal,
     totalAmount,
     itemCount,
@@ -145,7 +74,7 @@ function calcCart() {
   const $cartTotal = document.querySelector('#cart-total .text-2xl');
   if ($cartTotal) $cartTotal.textContent = `₩${Math.round(finalAmount).toLocaleString()}`;
 
-  updateStockInfo(productList);
+  $stockStatus.textContent = getStockInfoMessage(productList);
 
   renderBonusPoints(finalAmount, itemCount);
 }
@@ -325,48 +254,85 @@ function handleClickCartItem(event) {
 }
 
 function main() {
+  const elements = createElements();
+  setupEventListeners(elements);
+  appendElementsToDOM(elements);
+  initializeApp();
+}
+
+function createElements() {
   const $header = Header();
-
   const $selectorContainer = SelectorContainer();
-  $selectorContainer.appendChild($selector);
-  $selectorContainer.appendChild($addButton);
-  $selectorContainer.appendChild($stockStatus);
-
   const $leftColumn = LeftColumn();
-  $leftColumn.appendChild($selectorContainer);
-  $leftColumn.appendChild($cartItems);
-
   const $rightColumn = RightColumn();
-
   const $gridContainer = GridContainer();
-  $gridContainer.appendChild($leftColumn);
-  $gridContainer.appendChild($rightColumn);
-
   const $manualColumn = ManualColumn();
-
   const $manualOverlay = ManualOverlay();
+  const $manualToggle = ManualToggle();
+
+  return {
+    $header,
+    $selectorContainer,
+    $leftColumn,
+    $rightColumn,
+    $gridContainer,
+    $manualColumn,
+    $manualOverlay,
+    $manualToggle,
+  };
+}
+
+function setupEventListeners(elements) {
+  const { $manualOverlay, $manualColumn, $manualToggle } = elements;
+
   $manualOverlay.onclick = function (e) {
     if (e.target === $manualOverlay) {
       $manualOverlay.classList.add('hidden');
       $manualColumn.classList.add('translate-x-full');
     }
   };
-  $manualOverlay.appendChild($manualColumn);
 
-  const $manualToggle = ManualToggle();
   $manualToggle.onclick = function () {
     $manualOverlay.classList.toggle('hidden');
     $manualColumn.classList.toggle('translate-x-full');
   };
+}
 
+function appendElementsToDOM(elements) {
+  const {
+    $header,
+    $selectorContainer,
+    $leftColumn,
+    $rightColumn,
+    $gridContainer,
+    $manualColumn,
+    $manualOverlay,
+    $manualToggle,
+  } = elements;
+
+  // Build component hierarchy
+  $selectorContainer.appendChild($selector);
+  $selectorContainer.appendChild($addButton);
+  $selectorContainer.appendChild($stockStatus);
+
+  $leftColumn.appendChild($selectorContainer);
+  $leftColumn.appendChild($cartItems);
+
+  $gridContainer.appendChild($leftColumn);
+  $gridContainer.appendChild($rightColumn);
+
+  $manualOverlay.appendChild($manualColumn);
+
+  // Append to root
   const $root = document.getElementById('app');
   $root.appendChild($header);
   $root.appendChild($gridContainer);
   $root.appendChild($manualToggle);
   $root.appendChild($manualOverlay);
+}
 
+function initializeApp() {
   const productList = productStore.getProductList();
-
   renderSelectorOption(productList, $selector);
   calcCart();
 }
